refactor(app): extract useGameExecution hook (k4g)
Moves isGameRunning, screenshotCount, screenshot handlers, game running
IPC listeners, and auto-update notification effects out of App.tsx into
src/hooks/useGameExecution.ts.
