+++ src/App.tsx
@@ -58,6 +58,7 @@ import { useBlockManagement } from '@/hooks/useBlockManagement';
 import { useLoadingState } from '@/hooks/useLoadingState';
 import { useDirtyState } from '@/hooks/useDirtyState';
 import { useExternalFileChanges } from '@/hooks/useExternalFileChanges';
+import { useGameExecution } from '@/hooks/useGameExecution';
 import { formatErrorMessage } from '@/lib/formatErrorMessage';
 import { computeRouteCanvasLayout } from '@/lib/routeCanvasLayout';
 import { resolveWarpTarget } from '@/lib/warpTarget';
@@ -329,10 +330,6 @@ const App: React.FC = () => {
 
   const [nonRenpyWarningPath, setNonRenpyWarningPath] = useState<string | null>(null);
   
-  // --- State: Game Execution ---
-  const [isGameRunning, setIsGameRunning] = useState(false);
-  const [screenshotCount, setScreenshotCount] = useState(0);
-
   // --- State: Application and Project Settings ---
   const {
     appSettings,
@@ -1230,11 +1227,6 @@ const App: React.FC = () => {
     }
   }, [activeCanvasTabId, closeGoToLabelModal, setCenterOnBlockRequest, setCenterOnChoiceNodeRequest, setCenterOnRouteNodeRequest]);
 
-  const handleRunGame = useCallback(() => {
-    if (!window.electronAPI || !projectRootPath) return;
-    window.electronAPI.runGame(appSettings.renpyPath, projectRootPath);
-  }, [appSettings.renpyPath, projectRootPath]);
-
   const cleanupWarpTempFile = useCallback(async () => {
     if (!window.electronAPI || !projectRootPath) return;
 
@@ -1253,6 +1245,20 @@ const App: React.FC = () => {
       }
     }
   }, [projectRootPath]);
+
+  const {
+    isGameRunning,
+    screenshotCount,
+    handleRunGame,
+    handleOpenScreenshotsFolder,
+    handleClearScreenshots,
+    handleCopyLatestScreenshotPath,
+  } = useGameExecution({
+    projectRootPath,
+    renpyPath: appSettings.renpyPath,
+    addToast,
+    cleanupWarpTempFile,
+  });
 
   const resetWarpLaunchState = useCallback(() => {
     closeWarpVariablesModal();
@@ -1415,39 +1421,6 @@ const App: React.FC = () => {
       updateAppSettings(draft => { draft.isLeftSidebarOpen = true; });
     }
   }, [appSettings.isLeftSidebarOpen, updateAppSettings]);
-
-  // --- Screenshot Handlers ---
-  const refreshScreenshotCount = useCallback(async () => {
-    if (!window.electronAPI?.getScreenshotCount) return;
-    const count = await window.electronAPI.getScreenshotCount();
-    setScreenshotCount(count);
-  }, []);
-
-  // Note: Screenshot capture is now handled entirely in main process via global shortcut.
-
-  const handleOpenScreenshotsFolder = useCallback(async () => {
-    if (!window.electronAPI?.openScreenshotsFolder) return;
-    await window.electronAPI.openScreenshotsFolder();
-  }, []);
-
-  const handleClearScreenshots = useCallback(async () => {
-    if (!window.electronAPI?.clearScreenshots) return;
-    const result = await window.electronAPI.clearScreenshots();
-    if (result.success) {
-      addToast(`Cleared ${result.count} screenshot${result.count !== 1 ? 's' : ''}`, 'success');
-      await refreshScreenshotCount();
-      window.electronAPI.updateExplorerMenuState?.({ hasScreenshots: false });
-    }
-  }, [addToast, refreshScreenshotCount]);
-
-  const handleCopyLatestScreenshotPath = useCallback(async () => {
-    if (!window.electronAPI?.getLatestScreenshotPath) return;
-    const path = await window.electronAPI.getLatestScreenshotPath();
-    if (path) {
-      await navigator.clipboard.writeText(path);
-      addToast('Screenshot path copied to clipboard', 'success');
-    }
-  }, [addToast]);
 
   const {
       handleCreateNode, handleRenameNode, handleDeleteNode, handleMoveNode,
@@ -1622,70 +1595,6 @@ const App: React.FC = () => {
         });
         return removeListener;
   }, [handleNewProjectRequest, handleOpenProjectFolder, handleOpenWithRenpyCheck, loadProject, handleSaveAll, projectRootPath, appSettings.renpyPath, handleOpenStaticTab, handleToggleSearch, updateAppSettings, handleDeleteNode, explorerSelectedPaths, handleRefreshProject, handleOpenScreenshotsFolder, handleCloseTab, activePaneId, activeTabId, secondaryActiveTabId, openAboutModal, openSettingsModal, openShortcutsModal, openTutorial, setExplorerExternalAction]);
-
-  // --- Screenshot Count ---
-  useEffect(() => {
-    if (projectRootPath) {
-      void refreshScreenshotCount();
-    }
-  }, [projectRootPath, refreshScreenshotCount]);
-
-  // Listen for screenshot capture events from main process
-  useEffect(() => {
-    if (!window.electronAPI?.onScreenshotCaptured) return;
-    const removeListener = window.electronAPI.onScreenshotCaptured((data: { filename: string; filepath: string }) => {
-      // Show in-app toast (only works if renderer is alive)
-      addToast(`Screenshot saved: ${data.filename}`, 'success');
-      // Refresh count
-      void refreshScreenshotCount();
-    });
-    return removeListener;
-  }, [addToast, refreshScreenshotCount]);
-
-  // Update menu state when screenshot count changes
-  useEffect(() => {
-    if (window.electronAPI?.updateExplorerMenuState) {
-      window.electronAPI.updateExplorerMenuState({ hasScreenshots: screenshotCount > 0 });
-    }
-  }, [screenshotCount]);
-
-  // --- Game Running State ---
-  useEffect(() => {
-      if (!window.electronAPI) return;
-      const removeStarted = window.electronAPI.onGameStarted(() => setIsGameRunning(true));
-      const removeStopped = window.electronAPI.onGameStopped(() => {
-        setIsGameRunning(false);
-        void cleanupWarpTempFile();
-      });
-      const removeError = window.electronAPI.onGameError(() => {
-        setIsGameRunning(false);
-        void cleanupWarpTempFile();
-      });
-      return () => { removeStarted(); removeStopped(); removeError(); };
-  }, [cleanupWarpTempFile]);
-
-  // --- Auto-update notifications ---
-  useEffect(() => {
-      if (!window.electronAPI?.onUpdateAvailable) return;
-      const removeAvailable = window.electronAPI.onUpdateAvailable((version: string) => {
-          addToast(`Update v${version} is downloading in the background.`, 'info');
-      });
-      const removeNotAvailable = window.electronAPI.onUpdateNotAvailable?.(() => {
-          addToast("Vangard Studio is up to date.", 'info');
-      });
-      const removeError = window.electronAPI.onUpdateError?.(() => {
-          addToast('Could not check for updates. Check your connection and try again.', 'error');
-      });
-      const removeDownloaded = window.electronAPI.onUpdateDownloaded((version: string) => {
-          addToast(`Update v${version} ready — restart Vangard Studio to install.`, 'success');
-      });
-      return () => {
-          removeAvailable();
-          removeNotAvailable?.();
-          removeError?.();
-          removeDownloaded();
-      };
-  }, [addToast]);
 
   // --- Exit Handling ---
   const hasUnsavedSettingsRef = useRef(hasUnsavedSettings);
